make data error

+++ makeData.py
@@ -78,9 +78,9 @@ def kor_rim(naver_ann, naver_qdf):
     in_url = 'https://www.kisrating.com/ratingsStatistics/statics_spread.do'
     in_page = requests.get(in_url)
     in_tables = pd.read_html(in_page.text)
-    yeild = in_tables[0].iloc[-1,-1]
+    yeild = float(in_tables[0].iloc[-1,-1])
     # 네이버 ROE 추정치, 4분기 BPS 평균, 
-    BPS_mean = naver_qdf['BPS(원)'].rolling(4).mean()[-1]
+    BPS_mean = float(naver_qdf['BPS(원)'].rolling(4).mean()[-1])
     #rim = (BPS * ROE/ r) or (EPS / r)
-    rim_price = round(BPS_mean*naver_ann.iloc[-1,5] / yeild,2)
+    rim_price = round(BPS_mean*float(naver_ann.iloc[-1,5]) / yeild,2)
     return rim_price, yeild
